dynamic width and fadeup animation added to news landing
