+++ templates/news-landing/news-landing.css
@@ -17,7 +17,6 @@ main .filter-container .category-input {
     font-weight: 400;
     line-height: 1.01;
     outline: none;
-    min-width: 145px;
 }
 
 main .filter-container .category-dropdown {
@@ -167,7 +166,7 @@ main .news-listing {
     object-fit: cover;
 }
 
-.news-listing .news-item {
+.news-listing .news-item-wrapper {
     background-color: var(--news-bg-arbres);
     padding: 40px;
     grid-column: span 2;
@@ -175,12 +174,12 @@ main .news-listing {
     text-decoration: none;
 }
 
-.news-listing .news-item:hover {
+.news-listing .news-item-wrapper:hover {
     opacity: 0.7;
     cursor: pointer;
 }
 
-.news-listing .news-item.brown {
+.news-listing .news-item-wrapper.brown {
     background-color: var(--news-bg-bien);
 }
 
@@ -195,12 +194,12 @@ main .news-listing {
     letter-spacing: .06em;
 }
 
-.news-listing .news-item.horizontal-banner .news-category {
+.news-listing .news-item-wrapper.horizontal-banner .news-category {
     background-color: var(--site-color-2);
     color: var(--news-button-text-colour);
 }
 
-.news-listing .news-item.horizontal-banner h2 {
+.news-listing .news-item-wrapper.horizontal-banner h2 {
     color: var(--site-color);
     font-size: 40px;
     margin-top: 0;
@@ -216,7 +215,7 @@ main .news-listing {
     letter-spacing: .06em;
 }
 
-.news-listing .news-item.horizontal-banner .news-date {
+.news-listing .news-item-wrapper.horizontal-banner .news-date {
     color: var(--site-color);
 }
 
@@ -234,11 +233,11 @@ main .news-listing {
     letter-spacing: .02em;
 }
 
-.news-item .clear {
+.news-item-wrapper .clear {
     clear: both;
 }
 
-.news-listing .news-item.horizontal-banner {
+.news-listing .news-item-wrapper.horizontal-banner {
     background-color: var(--news-bg-enavant);
 }
 
@@ -253,15 +252,15 @@ main .news-listing {
         height: auto;
     }
 
-    .news-listing .news-item {
+    .news-listing .news-item-wrapper {
         grid-column: span 1;
     }
 
-    .news-listing .news-item.horizontal-banner {
+    .news-listing .news-item-wrapper.horizontal-banner {
         grid-column: span 2;
     }
 
-    .news-listing .news-item.horizontal-banner .news-image-wrapper {
+    .news-listing .news-item-wrapper.horizontal-banner .news-image-wrapper {
         width: calc((((100% - (12 * 30px)) / 12) * 6) + (5 * 30px));
         background-size: cover;
         position: relative;
@@ -269,7 +268,7 @@ main .news-listing {
         margin: 15px;
     }
 
-    .news-listing .news-item.horizontal-banner .news-content-wrapper {
+    .news-listing .news-item-wrapper.horizontal-banner .news-content-wrapper {
         width: calc((((100% - (12 * 30px)) / 12) * 6) + (5 * 30px));
         float: right;
         margin: 15px;
